Move search_data_gui to toolbar module
