+++ geemap/geemap.py
@@ -49,11 +49,11 @@ class Map(ipyleaflet.Map):
         if kwargs["ee_initialize"]:
             ee_initialize()
 
-        # Default map center location (lat, lon) and zoom level
-        latlon = [20, 0]
+        # Default map center center (lat, lon) and zoom level
+        center = [20, 0]
         zoom = 2
 
-        # Interchangeable parameters between ipyleaflet and folium
+        # Set map width and height
         if "height" not in kwargs.keys():
             kwargs["height"] = "600px"
         elif isinstance(kwargs["height"], int):
@@ -61,28 +61,33 @@ class Map(ipyleaflet.Map):
         if "width" in kwargs.keys() and isinstance(kwargs["width"], int):
             kwargs["width"] = str(kwargs["width"]) + "px"
 
+        # Convert folium params to ipyleaflet params
         if "location" in kwargs.keys():
             kwargs["center"] = kwargs["location"]
             kwargs.pop("location")
-        if "center" not in kwargs.keys():
-            kwargs["center"] = latlon
-
         if "zoom_start" in kwargs.keys():
             kwargs["zoom"] = kwargs["zoom_start"]
             kwargs.pop("zoom_start")
+
+        # Set map center and zoom level
+        if "center" not in kwargs.keys():
+            kwargs["center"] = center
         if "zoom" not in kwargs.keys():
             kwargs["zoom"] = zoom
         if "max_zoom" not in kwargs.keys():
             kwargs["max_zoom"] = 24
 
+        # Add Google Maps as the default basemap
         if "add_google_map" not in kwargs.keys() and "basemap" not in kwargs.keys():
             kwargs["add_google_map"] = True
+
+        # Enable scroll wheel zoom by default
         if "scroll_wheel_zoom" not in kwargs.keys():
             kwargs["scroll_wheel_zoom"] = True
 
+        # Whether to use lite mode, which contains only the zoom control.
         if "lite_mode" not in kwargs.keys():
             kwargs["lite_mode"] = False
-
         if kwargs["lite_mode"]:
             kwargs["data_ctrl"] = False
             kwargs["zoom_ctrl"] = True
@@ -95,6 +100,7 @@ class Map(ipyleaflet.Map):
             kwargs["toolbar_ctrl"] = False
             kwargs["attribution_ctrl"] = False
 
+        # Default controls to use
         if "data_ctrl" not in kwargs.keys():
             kwargs["data_ctrl"] = True
         if "zoom_ctrl" not in kwargs.keys():
@@ -118,10 +124,12 @@ class Map(ipyleaflet.Map):
         if "use_voila" not in kwargs.keys():
             kwargs["use_voila"] = False
 
+        # Use any basemap available through the basemap module, such as 'ROADMAP', 'OpenTopoMap'
         if "basemap" in kwargs:
             if isinstance(kwargs["basemap"], str):
                 kwargs["basemap"] = get_basemap(kwargs["basemap"])
 
+        # Disable some widgets if using Voila
         if os.environ.get("USE_VOILA") is not None:
             kwargs["use_voila"] = True
 
@@ -145,6 +153,7 @@ class Map(ipyleaflet.Map):
                 print("The sandbox path is invalid.")
                 self.sandbox_path = None
 
+        # Remove all default controls
         self.clear_controls()
 
         # The number of shapes drawn by the user using the DrawControl
@@ -164,6 +173,8 @@ class Map(ipyleaflet.Map):
 
         self.roi_start = False
         self.roi_end = False
+
+        # Default reducer to use
         if kwargs["ee_initialize"]:
             self.roi_reducer = ee.Reducer.mean()
         self.roi_reducer_scale = None
@@ -209,302 +220,9 @@ class Map(ipyleaflet.Map):
         self._expand_pixels = True
         self._expand_objects = False
 
-        # Adds search button and search box
-        search_button = widgets.ToggleButton(
-            value=False,
-            tooltip="Search location/data",
-            icon="globe",
-            layout=widgets.Layout(
-                width="28px", height="28px", padding="0px 0px 0px 4px"
-            ),
-        )
-
-        search_type = widgets.ToggleButtons(
-            options=["name/address", "lat-lon", "data"],
-            tooltips=[
-                "Search by place name or address",
-                "Search by lat-lon coordinates",
-                "Search Earth Engine data catalog",
-            ],
-        )
-        search_type.style.button_width = "110px"
-
-        search_box = widgets.Text(
-            placeholder="Search by place name or address",
-            tooltip="Search location",
-            layout=widgets.Layout(width="340px"),
-        )
-
-        search_output = widgets.Output(
-            layout={
-                "max_width": "340px",
-                "max_height": "350px",
-                "overflow": "scroll",
-            }
-        )
-
-        search_results = widgets.RadioButtons()
-
-        assets_dropdown = widgets.Dropdown(
-            options=[],
-            layout=widgets.Layout(min_width="279px", max_width="279px"),
-        )
-
-        import_btn = widgets.Button(
-            description="import",
-            button_style="primary",
-            tooltip="Click to import the selected asset",
-            layout=widgets.Layout(min_width="57px", max_width="57px"),
-        )
-
-        def get_ee_example(asset_id):
-            try:
-                import pkg_resources
-
-                pkg_dir = os.path.dirname(
-                    pkg_resources.resource_filename("geemap", "geemap.py")
-                )
-                with open(
-                    os.path.join(pkg_dir, "data/gee_f.json"), encoding="utf-8"
-                ) as f:
-                    functions = json.load(f)
-                details = [
-                    dataset["code"]
-                    for x in functions["examples"]
-                    for dataset in x["contents"]
-                    if x["name"] == "Datasets"
-                    if dataset["name"] == asset_id.replace("/", "_")
-                ]
-
-                return js_snippet_to_py(
-                    details[0],
-                    add_new_cell=False,
-                    import_ee=False,
-                    import_geemap=False,
-                    show_map=False,
-                )
-
-            except Exception as e:
-                pass
-            return
-
-        def import_btn_clicked(b):
-            if assets_dropdown.value is not None:
-                datasets = self.search_datasets
-                dataset = datasets[assets_dropdown.index]
-                id_ = dataset["id"]
-                code = get_ee_example(id_)
-
-                if not code:
-                    dataset_uid = "dataset_" + random_string(string_length=3)
-                    translate = {
-                        "image_collection": "ImageCollection",
-                        "image": "Image",
-                        "table": "FeatureCollection",
-                        "table_collection": "FeatureCollection",
-                    }
-                    datatype = translate[dataset["type"]]
-                    id_ = dataset["id"]
-                    line1 = "{} = ee.{}('{}')".format(dataset_uid, datatype, id_)
-                    action = {
-                        "image_collection": f"\nMap.addLayer({dataset_uid}, {{}}, '{id_}')",
-                        "image": f"\nMap.addLayer({dataset_uid}, {{}}, '{id_}')",
-                        "table": f"\nMap.addLayer({dataset_uid}, {{}}, '{id_}')",
-                        "table_collection": f"\nMap.addLayer({dataset_uid}, {{}}, '{id_}')",
-                    }
-                    line2 = action[dataset["type"]]
-                    code = [line1, line2]
-
-                contents = "".join(code).strip()
-                # create_code_cell(contents)
-                with search_output:
-                    search_output.clear_output(wait=True)
-                    print(
-                        "# The code has been copied to the clipboard. \n# Press Ctrl+V in a new cell to paste it.\n"
-                    )
-                    print(contents)
-
-        import_btn.on_click(import_btn_clicked)
-
-        html_widget = widgets.HTML()
-
-        def dropdown_change(change):
-            dropdown_index = assets_dropdown.index
-            if dropdown_index is not None and dropdown_index >= 0:
-                with search_output:
-                    search_output.clear_output(wait=True)
-                    print("Loading ...")
-                    datasets = self.search_datasets
-                    dataset = datasets[dropdown_index]
-                    dataset_html = ee_data_html(dataset)
-                    html_widget.value = dataset_html
-                    search_output.clear_output(wait=True)
-                    display(html_widget)
-
-        assets_dropdown.observe(dropdown_change, names="value")
-
-        assets_combo = widgets.HBox()
-        assets_combo.children = [import_btn, assets_dropdown]
-
-        def search_result_change(change):
-            result_index = search_results.index
-            locations = self.search_locations
-            location = locations[result_index]
-            latlon = (location.lat, location.lng)
-            self.search_loc_geom = ee.Geometry.Point(location.lng, location.lat)
-            marker = self.search_loc_marker
-            marker.location = latlon
-            self.center = latlon
-
-        search_results.observe(search_result_change, names="value")
-
-        def search_btn_click(change):
-            if change["new"]:
-                search_widget.children = [search_button, search_result_widget]
-                search_type.value = "name/address"
-            else:
-                search_widget.children = [search_button]
-                search_result_widget.children = [search_type, search_box]
-
-        search_button.observe(search_btn_click, "value")
-
-        def search_type_changed(change):
-            search_box.value = ""
-            search_output.clear_output()
-            if change["new"] == "data":
-                search_box.placeholder = (
-                    "Search GEE data catalog by keywords, e.g., elevation"
-                )
-                search_result_widget.children = [
-                    search_type,
-                    search_box,
-                    assets_combo,
-                    search_output,
-                ]
-            elif change["new"] == "lat-lon":
-                search_box.placeholder = "Search by lat-lon, e.g., 40, -100"
-                assets_dropdown.options = []
-                search_result_widget.children = [
-                    search_type,
-                    search_box,
-                    search_output,
-                ]
-            elif change["new"] == "name/address":
-                search_box.placeholder = "Search by place name or address, e.g., Paris"
-                assets_dropdown.options = []
-                search_result_widget.children = [
-                    search_type,
-                    search_box,
-                    search_output,
-                ]
-
-        search_type.observe(search_type_changed, names="value")
-
-        def search_box_callback(text):
-            if text.value != "":
-                if search_type.value == "name/address":
-                    g = geocode(text.value)
-                elif search_type.value == "lat-lon":
-                    g = geocode(text.value, reverse=True)
-                    if g is None and latlon_from_text(text.value):
-                        search_output.clear_output()
-                        latlon = latlon_from_text(text.value)
-                        self.search_loc_geom = ee.Geometry.Point(latlon[1], latlon[0])
-                        if self.search_loc_marker is None:
-                            marker = ipyleaflet.Marker(
-                                location=latlon,
-                                draggable=False,
-                                name="Search location",
-                            )
-                            self.search_loc_marker = marker
-                            self.add(marker)
-                            self.center = latlon
-                        else:
-                            marker = self.search_loc_marker
-                            marker.location = latlon
-                            self.center = latlon
-                        with search_output:
-                            print(f"No address found for {latlon}")
-                        return
-                elif search_type.value == "data":
-                    search_output.clear_output()
-                    with search_output:
-                        print("Searching ...")
-                    self.default_style = {"cursor": "wait"}
-                    ee_assets = search_ee_data(text.value, source="all")
-                    self.search_datasets = ee_assets
-                    asset_titles = [x["title"] for x in ee_assets]
-                    assets_dropdown.options = asset_titles
-                    search_output.clear_output()
-                    if len(ee_assets) > 0:
-                        assets_dropdown.index = 0
-                        html_widget.value = ee_data_html(ee_assets[0])
-                    else:
-                        html_widget.value = "No results found."
-                    with search_output:
-                        display(html_widget)
-                    self.default_style = {"cursor": "default"}
-
-                    return
-
-                self.search_locations = g
-                if g is not None and len(g) > 0:
-                    top_loc = g[0]
-                    latlon = (top_loc.lat, top_loc.lng)
-                    self.search_loc_geom = ee.Geometry.Point(top_loc.lng, top_loc.lat)
-                    if self.search_loc_marker is None:
-                        marker = ipyleaflet.Marker(
-                            location=latlon,
-                            draggable=False,
-                            name="Search location",
-                        )
-                        self.search_loc_marker = marker
-                        self.add(marker)
-                        self.center = latlon
-                    else:
-                        marker = self.search_loc_marker
-                        marker.location = latlon
-                        self.center = latlon
-                    search_results.options = [x.address for x in g]
-                    search_result_widget.children = [
-                        search_type,
-                        search_box,
-                        search_output,
-                    ]
-                    with search_output:
-                        search_output.clear_output(wait=True)
-                        display(search_results)
-                else:
-                    with search_output:
-                        search_output.clear_output()
-                        print("No results could be found.")
-
-        search_box.on_submit(search_box_callback)
-
-        search_result_widget = widgets.VBox([search_type, search_box])
-        search_widget = widgets.HBox([search_button])
-
-        search_event = ipyevents.Event(
-            source=search_widget, watched_events=["mouseenter", "mouseleave"]
-        )
-
-        def handle_search_event(event):
-            if event["type"] == "mouseenter":
-                search_widget.children = [search_button, search_result_widget]
-                # search_type.value = "name/address"
-            elif event["type"] == "mouseleave":
-                if not search_button.value:
-                    search_widget.children = [search_button]
-                    search_result_widget.children = [search_type, search_box]
-
-        search_event.on_dom_event(handle_search_event)
-
-        data_control = ipyleaflet.WidgetControl(
-            widget=search_widget, position="topleft"
-        )
-
         if kwargs.get("data_ctrl"):
-            self.add_control(control=data_control)
+            from .toolbar import search_data_gui
+            search_data_gui(self)
 
         search_marker = ipyleaflet.Marker(
             icon=ipyleaflet.AwesomeIcon(
@@ -865,11 +583,6 @@ class Map(ipyleaflet.Map):
             else:
                 tool = change["owner"]
                 tool_name = tools[tool.icon]["name"]
-                if tool_name == "to_image":
-                    tool_output.clear_output()
-                    save_map_widget.children = [save_type, file_chooser]
-                    if tool_output_control in self.controls:
-                        self.remove_control(tool_output_control)
                 if tool_name == "inspector":
                     inspector_output.clear_output()
                     self.inspector_checked = False
